deps/crypto: refactor crypto crate code structure
different submodules are split into different files and directories.
