+++ deps/crypto/src/native/aes256gcm.rs
@@ -1,0 +1,22 @@
+// Copyright (c) 2022 Alibaba Cloud
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+//! This mod implements aes-256-gcm encryption & decryption.
+
+use anyhow::*;
+use openssl::symm::Cipher;
+
+const TAG_LENGTH: usize = 16;
+
+pub fn decrypt(encrypted_data: &[u8], key: &[u8], iv: &[u8]) -> Result<Vec<u8>> {
+    let cipher = Cipher::aes_256_gcm();
+    if encrypted_data.len() < TAG_LENGTH {
+        bail!("Illegal length of ciphertext");
+    }
+
+    let (data, tag) = encrypted_data.split_at(encrypted_data.len() - TAG_LENGTH);
+    openssl::symm::decrypt_aead(cipher, key, Some(iv), &[], data, tag)
+        .map_err(|e| anyhow!(e.to_string()))
+}
